CASM-3552: Initial v0 api checker to scan for removed k8s apis in repos.
For now just running Pluto which requires valid yaml to work.

As such each user of this action will need to run helm template properly
based on each repos helm charts to a file.

If any deprecated or removed api's are found, a comment will be inserted
into the pull request.

The goal is to ensure any removed api usage blocks a pr until fixed, in
case new usages get inserted in a pr. And to comment on deprecated api
usage so that users are aware of upcoming k8s api removals.